+++ benches/highlight_utils/mod.rs
@@ -1,6 +1,5 @@
-use syntect::easy::HighlightLines;
-use syntect::highlighting::Theme;
-use syntect::parsing::{SyntaxReference, SyntaxSet};
+use syntect::highlighting::{HighlightIterator, HighlightState, Highlighter, Theme};
+use syntect::parsing::{ParseState, ScopeStack, SyntaxReference, SyntaxSet};
 
 /// Common helper for benchmarking highlighting.
 pub fn do_highlight(
@@ -9,10 +8,14 @@ pub fn do_highlight(
     syntax: &SyntaxReference,
     theme: &Theme,
 ) -> usize {
-    let mut h = HighlightLines::new(syntax, theme);
+    let highlighter = Highlighter::new(theme);
+    let mut highlight_state = HighlightState::new(&highlighter, ScopeStack::new());
+    let mut parse_state = ParseState::new(syntax);
     let mut count = 0;
     for line in s.lines() {
-        let regions = h.highlight_line(line, syntax_set).unwrap();
+        let ops = parse_state.parse_line(line, syntax_set).unwrap().ops;
+        let regions: Vec<_> =
+            HighlightIterator::new(&mut highlight_state, &ops[..], line, &highlighter).collect();
         count += regions.len();
     }
     count