+++ examples/latex-demo.rs
@@ -1,6 +1,5 @@
-use syntect::easy::HighlightLines;
-use syntect::highlighting::{Style, ThemeSet};
-use syntect::parsing::SyntaxSet;
+use syntect::highlighting::{HighlightIterator, HighlightState, Highlighter, Style, ThemeSet};
+use syntect::parsing::{ParseState, ScopeStack, SyntaxSet};
 use syntect::util::{as_latex_escaped, LinesWithEndings};
 
 fn main() {
@@ -11,10 +10,14 @@ fn main() {
     let syntax = ps.find_syntax_by_extension("rs").unwrap();
     let s = "pub struct Wow { hi: u64 }\nfn blah() -> u64 {}\n";
 
-    let mut h = HighlightLines::new(syntax, &ts.themes["InspiredGitHub"]);
+    let highlighter = Highlighter::new(&ts.themes["InspiredGitHub"]);
+    let mut highlight_state = HighlightState::new(&highlighter, ScopeStack::new());
+    let mut parse_state = ParseState::new(syntax);
     for line in LinesWithEndings::from(s) {
         // LinesWithEndings enables use of newlines mode
-        let ranges: Vec<(Style, &str)> = h.highlight_line(line, &ps).unwrap();
+        let ops = parse_state.parse_line(line, &ps).unwrap().ops;
+        let ranges: Vec<(Style, &str)> =
+            HighlightIterator::new(&mut highlight_state, &ops[..], line, &highlighter).collect();
         let escaped = as_latex_escaped(&ranges[..]);
         println!("\n{:?}", line);
         println!("\n{}", escaped);