+++ examples/syncat.rs
@@ -124,10 +124,8 @@ fn main() {
                 }
 
                 {
-                    let regions: Vec<(Style, &str)> = highlighter
-                        .highlight_lines
-                        .highlight_line(&line, &ss)
-                        .unwrap();
+                    let regions: Vec<(Style, &str)> =
+                        highlighter.highlight_line(&line, &ss).unwrap();
                     print!("{}", as_24_bit_terminal_escaped(&regions[..], true));
                 }
                 line.clear();